requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.10
python-dotenv>=1.0.0
graphviz>=0.20.1

# API Client Libraries
openai>=1.3.0
leptonai>=0.7.0

# LangChain integrations
langchain-core>=0.1.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Content processing
markdown>=3.4.0
